use unit16_t instead of int for pulse array
This saves 5152 byte (6.2% on ESP8266) of dynamic memory.

+++ src/pilight/libs/pilight/protocols/protocol.h
@@ -22,6 +22,7 @@
 //from ../core/options.h
 #define options_add(options, id, name, argtype, conftype, vartype, def, mask)
 
+#include <stdint.h>
 #include "../core/json.h"
 
 // from ../config/hardware.h
@@ -78,7 +79,7 @@ typedef struct protocol_t {
   unsigned long first;
   unsigned long second;
 
-  int *raw;
+  uint16_t *raw;
 
   hwtype_t hwtype;
   devtype_t devtype;
